requirements.txt
alembic==1.4.3
asyncpg==0.21.0
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
databases==0.4.1
dnspython==2.0.0
docstring-parser==0.7.3
elastic-apm==5.10.0
email-validator==1.1.2
Faker==5.0.2
fastapi==0.62.0
FastAPI-SQLAlchemy==0.2.1
h11==0.11.0
httpcore==0.12.2
httpx==0.16.1
httpx-oauth==0.3.4
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
makefun==1.9.5
Mako==1.1.3
MarkupSafe==1.1.1
minio==7.0.1
packaging==20.8
passlib==1.7.4
pika==1.1.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.20
starlette==0.13.6
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.3
zipp==3.4.0
